LayoutTests/editing/selection/move-by-word-visually-single-space-sigle-line-expected.txt


======== Move By Word ====
Test 1, LTR:
Move right by one word
"abc def hij opq rst"[0, 4, 8, 12, 16]
Move left by one word
"abc def hij opq rst"[19, 16, 12, 8, 4, 0]
Test 2, RTL:
Move left by one word
"abc def hij opq rst"[0, 15, 11, 7, 3]
Move right by one word
"abc def hij opq rst"[19, 3, 7, 11, 15, 0]
Test 3, LTR:
Move right by one word
"ZZZ QQQ BBB CCC XXX"[0, 15, 11, 7, 3]
Move left by one word
"ZZZ QQQ BBB CCC XXX"[19, 3, 7, 11, 15, 0]
Test 4, RTL:
Move left by one word
"ZZZ QQQ BBB CCC XXX"[0, 4, 8, 12, 16]
Move right by one word
"ZZZ QQQ BBB CCC XXX"[19, 16, 12, 8, 4, 0]
Test 5, LTR:
Move right by one word
"abc def hij AAA BBB WWW opr uvw xyz"[0, 4, 8, 12, 19, 15, 24, 28, 32]
Move left by one word
"abc def hij AAA BBB WWW opr uvw xyz"[35, 32, 28, 24, 15, 19, 12, 8, 4, 0]
Test 6, RTL:
Move left by one word
"abc def hij AAA BBB WWW opr uvw xyz"[0, 7, 3, 12, 16, 20, 24, 31, 27]
Move right by one word
"abc def hij AAA BBB WWW opr uvw xyz"[35, 27, 31, 24, 20, 16, 12, 3, 7, 0]
Test 7, LTR:
Move right by one word
"abc def ZQB RIG uvw xyz"[0, 4, 8, 11, 16, 20]
Move left by one word
"abc def ZQB RIG uvw xyz"[23, 20, 16, 11, 8, 4, 0]
Test 8, RTL:
Move left by one word
"abc def ZQB RIG uvw xyz"[0, 3, 8, 12, 16, 19]
Move right by one word
"abc def ZQB RIG uvw xyz"[23, 19, 16, 12, 8, 3, 0]
Test 9, LTR:
Move right by one word
"aaa AAA bbb"[0, 4, 8]
Move left by one word
"aaa AAA bbb"[11, 8, 4, 0]
Test 10, RTL:
Move left by one word
"aaa AAA bbb"[0, 4, 8, 11]
Move right by one word
"aaa AAA bbb"[11, 8, 4, 0]
Test 11, LTR:
Move right by one word
"AAA BBB WWW aaa bbb ccc DDD SSS UUU"[0, 7, 3, 12, 16, 20, 24, 31, 27]
Move left by one word
"AAA BBB WWW aaa bbb ccc DDD SSS UUU"[35, 27, 31, 24, 20, 16, 12, 3, 7, 0]
Test 12, RTL:
Move left by one word
"AAA BBB WWW aaa bbb ccc DDD SSS UUU"[0, 4, 8, 12, 19, 15, 24, 28, 32]
Move right by one word
"AAA BBB WWW aaa bbb ccc DDD SSS UUU"[35, 32, 28, 24, 15, 19, 12, 8, 4, 0]
Test 13, LTR:
Move right by one word
"AAA BBB aaa bbb WWW DDD"[0, 3, 8, 12, 16, 19]
Move left by one word
"AAA BBB aaa bbb WWW DDD"[23, 19, 16, 12, 8, 3, 0]
Test 14, RTL:
Move left by one word
"AAA BBB aaa bbb WWW DDD"[0, 4, 8, 11, 16, 20]
Move right by one word
"AAA BBB aaa bbb WWW DDD"[23, 20, 16, 11, 8, 4, 0]
Test 15, LTR:
Move right by one word
"ZQB abc RIG"[0, 4, 8, 11]
Move left by one word
"ZQB abc RIG"[11, 8, 4, 0]
Test 16, RTL:
Move left by one word
"ZQB abc RIG"[0, 4, 8]
Move right by one word
"ZQB abc RIG"[11, 8, 4, 0]
Test 17, LTR:
Move right by one word
"AAA "[0, 3]
Move left by one word
"BB"[2], "AAA "[3, 0]
Test 18, RTL:
Move left by one word
"AAA "[0, 4]
Move right by one word
"BB"[2], "AAA "[4, 0]
Test 19, LTR:
Move right by one word
"abc def "[0, 4, 8], "hij opq"[4], " rst uvw"[1, 5]
Move left by one word
" rst uvw"[8, 5, 1], "hij opq"[4], "abc def "[8, 4, 0]
Test 20, RTL:
Move left by one word
"abc def "[0], " rst uvw"[4], "hij opq"[3], "abc def "[7, 3]    FAIL expected: ["abc def "[ 0, ]" rst uvw"[ 4, ]"hij opq"[ 7,  3, ]"abc def "[ 7,  3]
" rst uvw"[4], "hij opq"[3]   FAIL expected "hij opq"[ 7]
" rst uvw"[3], "hij opq"[3]   FAIL expected "hij opq"[ 7]
" rst uvw"[2], "hij opq"[3]   FAIL expected "hij opq"[ 7]
" rst uvw"[1], "hij opq"[3]   FAIL expected "hij opq"[ 7]
Move right by one word
" rst uvw"[8], "abc def "[3, 7], "hij opq"[3], " rst uvw"[4], "abc def "[0]    FAIL expected: [" rst uvw"[ 8, ]"abc def "[ 3,  7, ]"hij opq"[ 3,  7, ]" rst uvw"[ 4, ]"abc def "[ 0]
"hij opq"[3], " rst uvw"[4]   FAIL expected "hij opq"[ 7]
"hij opq"[4], " rst uvw"[4]   FAIL expected "hij opq"[ 7]
"hij opq"[5], " rst uvw"[4]   FAIL expected "hij opq"[ 7]
"hij opq"[6], " rst uvw"[4]   FAIL expected "hij opq"[ 7]
Test 21, RTL:
Move left by one word
"abc def "[0], " rst uvw"[4], "hij opq"[3], "abc def "[7, 3]    FAIL expected: ["abc def "[ 0, ]" rst uvw"[ 4, ]"hij opq"[ 7,  3, ]"abc def "[ 7,  3]
" rst uvw"[4], "hij opq"[3]   FAIL expected "hij opq"[ 7]
" rst uvw"[3], "hij opq"[3]   FAIL expected "hij opq"[ 7]
" rst uvw"[2], "hij opq"[3]   FAIL expected "hij opq"[ 7]
" rst uvw"[1], "hij opq"[3]   FAIL expected "hij opq"[ 7]
Move right by one word
" rst uvw"[8], "abc def "[3, 7], "hij opq"[3], " rst uvw"[4], "abc def "[0]    FAIL expected: [" rst uvw"[ 8, ]"abc def "[ 3,  7, ]"hij opq"[ 3,  7, ]" rst uvw"[ 4, ]"abc def "[ 0]
"hij opq"[3], " rst uvw"[4]   FAIL expected "hij opq"[ 7]
"hij opq"[4], " rst uvw"[4]   FAIL expected "hij opq"[ 7]
"hij opq"[5], " rst uvw"[4]   FAIL expected "hij opq"[ 7]
"hij opq"[6], " rst uvw"[4]   FAIL expected "hij opq"[ 7]
Test 22, LTR:
Move right by one word
"abc def "[0, 4, 8], "hij opq"[4], " rst uvw"[1, 5]
Move left by one word
" rst uvw"[8, 5, 1], "hij opq"[4], "abc def "[8, 4, 0]
Test 23, RTL:
Move left by one word
"ABD DSU "[0, 4, 8], "EJH FUX"[4], "FFZ LIG"[4]
Move right by one word
"FFZ LIG"[7, 4], "EJH FUX"[4], "ABD DSU "[8, 4, 0]
Test 24, LTR:
Move right by one word
"ABD DSU "[0], "FFZ LIG"[3], "EJH FUX"[3], "ABD DSU "[7, 3]
Move left by one word
"FFZ LIG"[7], "ABD DSU "[3, 7], "EJH FUX"[3], "FFZ LIG"[3], "ABD DSU "[0]
Test 25, RTL:
Move left by one word
"ABD DSU "[0, 4, 8], "EJH FUX"[4], "FFZ LIG"[4]
Move right by one word
"FFZ LIG"[7, 4], "EJH FUX"[4], "ABD DSU "[8, 4, 0]
Test 26, LTR:
Move right by one word
"ABD DSU "[0], "FFZ LIG"[3], "EJH FUX"[3], "ABD DSU "[7, 3]
Move left by one word
"FFZ LIG"[7], "ABD DSU "[3, 7], "EJH FUX"[3], "FFZ LIG"[3], "ABD DSU "[0]
Test 27, RTL:
Move left by one word
"ABD DSU "[0, 4, 8], "abc def"[3], "FFZ LIG"[4]
Move right by one word
"FFZ LIG"[7, 4], "abc def"[3], "ABD DSU "[8, 4, 0]
Test 28, LTR:
Move right by one word
"ABD DSU "[0], "FFZ LIG"[3], "ABD DSU "[8], "abc def"[4], "ABD DSU "[7, 3]
Move left by one word
"FFZ LIG"[7], "ABD DSU "[3, 7], "abc def"[4], "ABD DSU "[8], "FFZ LIG"[3]    FAIL expected: ["FFZ LIG"[ 7, ]"ABD DSU "[ 3,  7, ]"abc def"[ 4, ]"ABD DSU "[ 8, ]"FFZ LIG"[ 3, ]"ABD DSU "[ 0]
"FFZ LIG"[3, 3]   FAIL expected "ABD DSU "[ 0]
"FFZ LIG"[4, 4]   FAIL expected "ABD DSU "[ 0]
"FFZ LIG"[5, 5]   FAIL expected "ABD DSU "[ 0]
"FFZ LIG"[6, 6]   FAIL expected "ABD DSU "[ 0]
Test 29, RTL:
Move left by one word
"ABD DSU "[0, 4, 8], "abc def"[3], "FFZ LIG"[4]
Move right by one word
"FFZ LIG"[7, 4], "abc def"[3], "ABD DSU "[8, 4, 0]
Test 30, LTR:
Move right by one word
"ABD DSU "[0, 3, 8], "abc def"[4], "FFZ LIG"[3]
Move left by one word
"FFZ LIG"[7, 3], "abc def"[4], "ABD DSU "[8, 3, 0]
Test 31, RTL:
Move left by one word
"ABD opq DSU "[0, 4, 8, 12], "abc AAA def"[8, 4, 3], "FFZ rst LIG"[4, 8]    FAIL expected: ["ABD opq DSU "[ 0,  4,  8,  12, ]"abc AAA def"[ 4,  3, ]"FFZ rst LIG"[ 4,  8]
"ABD opq DSU "[12], "abc AAA def"[8]   FAIL expected "abc AAA def"[ 4]
"abc AAA def"[10, 8]   FAIL expected "abc AAA def"[ 4]
"abc AAA def"[9, 8]   FAIL expected "abc AAA def"[ 4]
Move right by one word
"FFZ rst LIG"[11, 8, 4], "abc AAA def"[3, 4, 8], "ABD opq DSU "[12, 8, 4, 0]    FAIL expected: ["FFZ rst LIG"[ 11,  8,  4, ]"abc AAA def"[ 3,  4, ]"ABD opq DSU "[ 12,  8,  4,  0]
"abc AAA def"[4, 8]   FAIL expected "ABD opq DSU "[ 12]
Test 32, LTR:
Move right by one word
"ABD opq DSU "[0, 4], "abc AAA def"[8, 4], "ABD opq DSU "[12, 11], "FFZ rst LIG"[4, 8, 11]    FAIL expected: ["ABD opq DSU "[ 0,  4,  8, ]"abc AAA def"[ 8,  7, ]"ABD opq DSU "[ 12,  11, ]"FFZ rst LIG"[ 4,  8,  11]
"ABD opq DSU "[4], "abc AAA def"[8]   FAIL expected "ABD opq DSU "[ 8]
"ABD opq DSU "[5], "abc AAA def"[8]   FAIL expected "ABD opq DSU "[ 8]
"ABD opq DSU "[6], "abc AAA def"[8]   FAIL expected "ABD opq DSU "[ 8]
"ABD opq DSU "[7], "abc AAA def"[8]   FAIL expected "ABD opq DSU "[ 8]
"abc AAA def"[8, 4]   FAIL expected "abc AAA def"[ 7]
"abc AAA def"[9, 4]   FAIL expected "abc AAA def"[ 7]
"abc AAA def"[10, 4]   FAIL expected "abc AAA def"[ 7]
"abc AAA def"[11, 4]   FAIL expected "abc AAA def"[ 7]
"abc AAA def"[7, 4]   FAIL expected "ABD opq DSU "[ 12]
"abc AAA def"[6, 4]   FAIL expected "ABD opq DSU "[ 12]
"abc AAA def"[5, 4]   FAIL expected "ABD opq DSU "[ 12]
Move left by one word
"FFZ rst LIG"[11, 8, 4], "ABD opq DSU "[11, 12], "abc AAA def"[7, 8], "ABD opq DSU "[4, 0]    FAIL expected: ["FFZ rst LIG"[ 11,  8,  4, ]"ABD opq DSU "[ 11,  12, ]"abc AAA def"[ 7,  8, ]"ABD opq DSU "[ 8,  4,  0]
"abc AAA def"[8], "ABD opq DSU "[4]   FAIL expected "ABD opq DSU "[ 8]
"FFZ rst LIG"[1], "ABD opq DSU "[4]   FAIL expected "ABD opq DSU "[ 8]
"FFZ rst LIG"[2], "ABD opq DSU "[4]   FAIL expected "ABD opq DSU "[ 8]
Test 33, RTL:
Move left by one word
"ABD opq DSU "[0, 4, 8, 12], "abc AAA def"[4, 8], "FFZ rst LIG"[4, 8]
Move right by one word
"FFZ rst LIG"[11, 8, 4], "abc AAA def"[8, 4], "ABD opq DSU "[12, 8, 4, 0]
Test 34, LTR:
Move right by one word
"ABD opq DSU "[0, 4, 8, 12], "abc AAA def"[4, 8], "FFZ rst LIG"[4, 8, 11]
Move left by one word
"FFZ rst LIG"[11, 8, 4], "abc AAA def"[8, 4], "ABD opq DSU "[12, 8, 4, 0]
Test 35, LTR:
Move right by one word
"aaa "[0, 4], "bbb AAA "[4, 7]
Move left by one word
"FFZ"[3], "bbb AAA "[7, 4], "aaa "[4, 0]
Test 36, RTL:
Move left by one word
"ABD opq rst DSU "[0, 4, 7, 12, 16], "abc uvw AAA def lmn"[12, 8, 7, 3], "ABW hij xyz FXX"[4, 7, 12]    FAIL expected: ["ABD opq rst DSU "[ 0,  4,  7,  12,  16, ]"abc uvw AAA def lmn"[ 15,  8,  7,  3, ]"ABW hij xyz FXX"[ 4,  7,  12]
"ABD opq rst DSU "[16], "abc uvw AAA def lmn"[12]   FAIL expected "abc uvw AAA def lmn"[ 15]
"abc uvw AAA def lmn"[18, 12]   FAIL expected "abc uvw AAA def lmn"[ 15]
"abc uvw AAA def lmn"[17, 12]   FAIL expected "abc uvw AAA def lmn"[ 15]
"abc uvw AAA def lmn"[16, 12]   FAIL expected "abc uvw AAA def lmn"[ 15]
"abc uvw AAA def lmn"[15, 12]   FAIL expected "abc uvw AAA def lmn"[ 8]
"abc uvw AAA def lmn"[14, 12]   FAIL expected "abc uvw AAA def lmn"[ 8]
"abc uvw AAA def lmn"[13, 12]   FAIL expected "abc uvw AAA def lmn"[ 8]
Move right by one word
"ABW hij xyz FXX"[15, 12, 7, 4], "abc uvw AAA def lmn"[3, 7, 8, 15], "ABD opq rst DSU "[16, 12, 7, 4, 0]
